mysite/requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.2.7
django-import-export==4.3.10
django-tinymce==4.1.0
djangorestframework==3.16.1
pillow==12.0.0
sqlparse==0.5.3
tablib==3.9.0
